FMS emulation: force the negation of x to happen _before_ the multiply
Recent clang has started to emit a single `fnmul` instruction for `-x * y`, the semantics of which are a multiply followed by a negate. We instead want a negate followed by a multiply, which is ever so slightly different when FPCR has the DN bit set. The `fmsub` instruction does a negate followed by a multiply-add, and setting z to -0 reduces that multiply-add to just a multiply, giving us what we want.